changed a bunch of dictionaries to use floats instead of ints

+++ SecretProject/SecretProject/Class/CollisionDetection/Collider.cs
@@ -73,7 +73,7 @@ namespace SecretProject.Class.CollisionDetection
 
 
 
-        public bool DidCollide(Dictionary<int,ObjectBody> objectBody, Vector2 position)
+        public bool DidCollide(Dictionary<float,ObjectBody> objectBody, Vector2 position)
         {
             
             foreach (var body in objectBody.Values)
