+++ SecretProject/SecretProject/Class/StageFolder/Interior.cs
@@ -56,9 +56,9 @@ namespace SecretProject.Class.StageFolder
         public string StageName { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         public List<Crop> AllCrops { get; set; }
         public bool IsLoaded { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-        Dictionary<int, ObjectBody> ILocation.AllObjects { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+        Dictionary<float, ObjectBody> ILocation.AllObjects { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
 
-        Dictionary<int, Crop> ILocation.AllCrops { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+        Dictionary<float, Crop> ILocation.AllCrops { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         ITileManager ILocation.AllTiles { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
 
         public void Draw(GraphicsDevice graphics, RenderTarget2D mainTarget, RenderTarget2D lightsTarget,GameTime gameTime, SpriteBatch spriteBatch, MouseManager mouse, Player player)
